demo: Fetch tiles and metadata over HTTPS
If GitHub Pages starts supporting HTTPS for custom domains, we'll need
to fetch everything over HTTPS to avoid mixed-content warnings in the
browser.  The S3 website endpoint doesn't support HTTPS.  The regular
endpoint only supports HTTPS with an AWS wildcard cert, and as a result
the bucket name can only contain dots if old-style path-based hosting
is used.  Rename the bucket, switch to HTTPS using the regular endpoint
with virtual hosting, and drop complications that were needed for the
website endpoint.