index: link to paper PDF on jpathinformatics.org
We've been rehosting the PDF on download.openslide.org because
jpathinformatics.org requires clicking through a landing page.  However,
the PDF is the last user of download.openslide.org, and it'd be nice to
fully deprecate that service, which is HTTP-only.  We could rehost the
PDF in this repo (it's 1.8 MB), but instead take the path of least
resistance and link to the official site.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>