Document pre-commit checks
The overlarge test case check was also moved to pre-commit and is
mentioned in docs/testsuite, but leave that doc as is.  The problem
isn't if a large test case is committed to a fork, but if it's merged to
the main repo.  Talking about pre-commit checks there woould likely
confuse the issue.

See: https://github.com/openslide/openslide/pull/540
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>