synctiles: don't set explicit public-read ACL on objects
We're moving to a bucket with per-object ACLs disabled.  This is simpler
to administer but requires the bucket to be manually configured with a
policy allowing public reads.  synctiles does not set a bucket policy
automatically, as it does with CORS, to avoid clobbering any nuanced
policy configured by the bucket owner.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>